pyproject.toml
[project]
name = "dauntless"
version = "0.1.0"
dependencies = [
    "numpy>=2.2.6",
    "pillow>=11.3.0",
]
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "matplotlib>=3.10.3",
    "opencv-python>=4.12.0.88",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/dauntless"]

[tool.pyright]
venvPath = "."
venv = ".venv"
